user-management/requirements.txt
Flask==2.3.3                 # Web framework
mysql-connector-python==8.0.33 # MySQL database connector
bcrypt==4.0.1               # Password hashing
